+++ R/app_config.R
@@ -6,7 +6,8 @@ save_user_config <- function(code_style,
                              model,
                              custom_prompt,
                              stream,
-                             read_docs) {
+                             read_docs,
+                             audio_input) {
   if (is.null(custom_prompt)) custom_prompt <- ""
   config <-
     data.frame(
@@ -18,7 +19,8 @@ save_user_config <- function(code_style,
       model,
       custom_prompt,
       stream,
-      read_docs
+      read_docs,
+      audio_input
     )
   user_config_path <- tools::R_user_dir("gptstudio", which = "config")
   user_config <- file.path(user_config_path, "config.yml")
@@ -40,7 +42,8 @@ set_user_options <- function(config) {
     gptstudio.custom_prompt = config$custom_prompt,
     gptstudio.stream        = config$stream,
     # added in v.3.1+ dev version
-    gptstudio.read_docs     = config$read_docs
+    gptstudio.read_docs     = config$read_docs,
+    gptstudio.audio_input   = config$audio_input
   )
   options(op_gptstudio)
   invisible()