+++ R/gptstudio-package.R
@@ -5,7 +5,6 @@
 #' @import cli
 #' @import rlang
 #' @import httr2
-#' @importFrom assertthat assert_that is.string is.number is.count
 #' @importFrom glue glue
 ## gptstudio namespace: end
 NULL