+++ R/mod_app.R
@@ -1,16 +1,18 @@
 #' App UI
 #'
 #' @param id id of the module
-#' @inheritParams run_chatgpt_app
+#' @inheritParams gptstudio_run_chat_app
 #'
 #' @import htmltools
 #' @import shiny
 #'
-mod_app_ui <- function(id, ide_colors = get_ide_theme_info()) {
+mod_app_ui <- function(id,
+                       ide_colors = get_ide_theme_info(),
+                       code_theme_url = get_highlightjs_theme()) {
   ns <- NS(id)
   translator <- create_translator(language = getOption("gptstudio.language"))
   tagList(
-    waiter::use_waiter(),
+    useBusyIndicators(),
     bslib::page_fluid(
       theme = create_chat_app_theme(ide_colors),
       title = "ChatGPT from gptstudio",
@@ -29,8 +31,7 @@ mod_app_ui <- function(id, ide_colors = get_ide_theme_info()) {
           class = "row justify-content-center h-100",
           div(
             class = "col h-100",
-            style = htmltools::css(`max-width` = "800px"),
-            mod_chat_ui(ns("chat"), translator)
+            mod_chat_ui(ns("chat"), translator, code_theme_url)
           )
         )
       )
@@ -41,7 +42,7 @@ mod_app_ui <- function(id, ide_colors = get_ide_theme_info()) {
 #' App Server
 #'
 #' @inheritParams mod_app_ui
-#' @inheritParams run_chatgpt_app
+#' @inheritParams gptstudio_run_chat_app
 #'
 mod_app_server <- function(id, ide_colors = get_ide_theme_info()) {
   moduleServer(id, function(input, output, session) {
@@ -63,6 +64,7 @@ mod_app_server <- function(id, ide_colors = get_ide_theme_info()) {
 #'
 #' @return hex color
 rgb_str_to_hex <- function(rgb_string) {
+  check_installed("grDevices")
   rgb_vec <- unlist(strsplit(gsub("[rgba() ]", "", rgb_string), ","))
   grDevices::rgb(
     red = as.numeric(rgb_vec[1]),
@@ -78,12 +80,13 @@ rgb_str_to_hex <- function(rgb_string) {
 #'
 #' Create a bslib theme that matches the user's RStudio IDE theme.
 #'
-#' @inheritParams run_chatgpt_app
+#' @inheritParams gptstudio_run_chat_app
 #'
 #' @return A bslib theme
 create_chat_app_theme <- function(ide_colors = get_ide_theme_info()) {
   bslib::bs_theme(
     version = 5,
+    preset = "shiny",
     bg = ide_colors$bg,
     fg = ide_colors$fg,
     font_scale = 0.9,
@@ -93,38 +96,56 @@ create_chat_app_theme <- function(ide_colors = get_ide_theme_info()) {
 }
 
 
-#' Get IDE theme information.
+#' Get IDE Theme Information
 #'
-#' This function returns a list with the current IDE theme's information.
+#' Retrieves the current RStudio IDE theme information including whether it is a dark theme,
+#' and the background and foreground colors in hexadecimal format.
 #'
-#' @return A list with three components:
-#' \item{is_dark}{A boolean indicating whether the current IDE theme is dark.}
-#' \item{bg}{The current IDE theme's background color.}
-#' \item{fg}{The current IDE theme's foreground color.}
+#' @return A list with the following components:
+#'   \item{is_dark}{A logical indicating whether the current IDE theme is dark.}
+#'   \item{bg}{A character string representing the background color of the IDE theme in hex format.}
+#'   \item{fg}{A character string representing the foreground color of the IDE theme in hex format.}
+#'
+#' If RStudio is unavailable, returns the fallback theme details.
+#'
+#' @examples
+#' theme_info <- get_ide_theme_info()
+#' print(theme_info)
 #'
 #' @export
-#'
 get_ide_theme_info <- function() {
   if (rstudioapi::isAvailable()) {
-    rstudio_theme_info <- rstudioapi::getThemeInfo()
+    tryCatch(
+      {
+        rstudio_theme_info <- rstudioapi::getThemeInfo()
 
-    # create a list with three components
-    list(
-      is_dark = rstudio_theme_info$dark,
-      bg = rgb_str_to_hex(rstudio_theme_info$background),
-      fg = rgb_str_to_hex(rstudio_theme_info$foreground)
+        # Create a list with theme components
+        list(
+          is_dark = rstudio_theme_info$dark,
+          bg = rgb_str_to_hex(rstudio_theme_info$background),
+          fg = rgb_str_to_hex(rstudio_theme_info$foreground)
+        )
+      },
+      error = function(e) {
+        cli::cli_warn("Error fetching theme info from RStudio: {e$message}")
+        fallback_theme()
+      }
     )
   } else {
-    if (interactive()) cli::cli_inform("Using fallback ide theme")
-
-    # create a list with three components with fallback values
-    list(
-      is_dark = TRUE,
-      bg = "#002B36",
-      fg = "#93A1A1"
-    )
+    if (interactive()) cli::cli_inform("RStudio is not available. Using fallback IDE theme.")
+    fallback_theme()
   }
 }
+
+# Fallback function to provide default values
+fallback_theme <- function() {
+  list(
+    is_dark = TRUE,
+    bg = "#181818",
+    fg = "#C1C1C1"
+  )
+}
+
 
 html_dependencies <- function() {
   htmltools::htmlDependency(