+++ R/run_chatgpt_app.R
@@ -4,14 +4,17 @@
 #' script.
 #'
 #' @param ide_colors List containing the colors of the IDE theme.
+#' @param code_theme_url URL to the highlight.js theme
 #' @inheritParams shiny::runApp
 #'
 #' @return Nothing.
 #' @export
-run_chatgpt_app <- function(ide_colors = get_ide_theme_info(),
-                            host = getOption("shiny.host", "127.0.0.1"),
-                            port = getOption("shiny.port")) {
-  ui <- mod_app_ui("app", ide_colors)
+gptstudio_run_chat_app <- function(ide_colors = get_ide_theme_info(),
+                                   code_theme_url = get_highlightjs_theme(),
+                                   host = getOption("shiny.host", "127.0.0.1"),
+                                   port = getOption("shiny.port")) {
+  check_installed("future")
+  ui <- mod_app_ui("app", ide_colors, code_theme_url)
 
   server <- function(input, output, session) {
     mod_app_server("app", ide_colors)