+++ NEWS.md
@@ -4,6 +4,13 @@
 - Added claude-3.5-sonnet model from Anthropic.
 - Set gpt-4o-mini as default model for OpenAI. #219
 - Fixed bugs with Azure OpenAI service. #223
+- Add audio input option for chat app. #224
+- Fix bug with chat app not loading on linux. #224
+- Allow chat app to run in Positron (not yet as background job) #224
+- API calls now run async with ExtendedTask. #224
+- New styling of chat app. #224
+- Add code syntax highlighting to chat app. #224
+- Replace curl calls with httr2. #224
 
 ## gptstudio 0.4.0